requirements.txt
mcp
openai
python-dotenv
requests
uvicorn
starlette
sse_starlette
